Add files via upload
Heart failure (HF) is a pathophysiological state in which the heart fails to supply required amount of blood and oxygen to the body. HF is widespread all over the world, being the leading cause of both hospitalization and readmission amongst older adults (≥ 65).
The dataset is provided by Lombardy Region: administrative healthcare databases are created from records of service utilization and payments for payer or hospital billing 
purposes. Such databases contain large amounts of information on the diagnoses, procedures, resource utilization, and costs or charges of thousands (even millions) of patients.
Our purpose is to:
1. Identify groups of patients with different characteristics
2. Analyse the additional contribution of the hospital-level features across clusters
3. Study if there are differences between clusters of patients that were cured in different hospitals (effect of the hospital) using techniques as mixed effects models